student_performance/requirements.txt
Django>=4.0
djangorestframework
pandas
numpy
scikit-learn
tensorflow>=2.10
joblib